fix: codeql issues

+++ src/OpenFga.Sdk/Model/ReadRequest.cs
@@ -189,13 +189,13 @@ namespace OpenFga.Sdk.Model
         public IEnumerable<System.ComponentModel.DataAnnotations.ValidationResult> Validate(ValidationContext validationContext)
         {
             // PageSize (int) maximum
-            if (this.PageSize > (int)100)
+            if (this.PageSize > 100)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for PageSize, must be a value less than or equal to 100.", new [] { "PageSize" });
             }
 
             // PageSize (int) minimum
-            if (this.PageSize < (int)1)
+            if (this.PageSize < 1)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for PageSize, must be a value greater than or equal to 1.", new [] { "PageSize" });
             }
